+++ src/Ordering.ts
@@ -1,7 +1,7 @@
 import { Setoid } from './Setoid'
 import { Semigroup } from './Semigroup'
 
-export type Ordering = 'LT' | 'EQ' | 'GT';
+export type Ordering = 'LT' | 'EQ' | 'GT'
 
 export const orderingSetoid: Setoid<Ordering> = {
   equals(a, b) {