+++ src/Traversable.ts
@@ -5,7 +5,7 @@ import { Applicative } from './Applicative'
 import { identity, Function1 } from './function'
 
 export interface Traversable<T> extends Functor<T>, Foldable<T> {
-  traverse<F, A, B>(applicative: Applicative<F>, f: Function1<A, HKT<F, B>>, ta: HKT<T, A>): HKT<F, HKT<T, B>>;
+  traverse<F, A, B>(applicative: Applicative<F>, f: Function1<A, HKT<F, B>>, ta: HKT<T, A>): HKT<F, HKT<T, B>>
 }
 
 export function sequence<F, T, A>(